Need to read in dimensions into 32-bit structure
The current codebase tries to read 4 ints (16 bytes) directly into Latt.dims, but Latt.dims is an array of 4 size_t objects (32 bytes), which is too large to hold them.  In the current implementation, the x- and y-dimensions are read as being several billion, and z- and t-dimensions are both zero, so reading a MILC configuration fails.  By storing into ints first, we can read the dimensions correctly and then copy these into the Latt.dims.